+++ src/pages/Problems/index.ts
@@ -1,0 +1,1 @@
+export { Problems } from './Problems';
